Moves methods to Wrongs module
